sim_record/requirements.txt
mujoco
onnxruntime
numpy
matplotlib
pyyaml
